+++ ddrescuestatus.h
@@ -8,13 +8,13 @@ class DdrescueStatus
 {
 private:
     bool m_parseOk;
-    long m_rescued;
-    long m_errsize;
+    long long m_rescued;
+    long long m_errsize;
     long m_currentRate;
-    long m_ipos;
+    long long m_ipos;
     long m_errors;
     long m_averageRate;
-    long m_opos;
+    long long m_opos;
     int m_timeFromLastSuccessfulRead;
 
     QString m_status;
@@ -25,13 +25,13 @@ public:
     bool parse(QString const & txt);
 
     bool parseOk() const;
-    long rescued() const;
-    long errsize() const;
+    long long rescued() const;
+    long long errsize() const;
     long currentRate() const;
-    long ipos() const;
+    long long ipos() const;
     long errors() const;
     long averageRate() const;
-    long opos() const;
+    long long opos() const;
     int timeFromLastSuccessfulRead() const;
     QString status() const;
